backend/requirements.txt
# Core API
fastapi>=0.115.0
uvicorn>=0.30.6
python-multipart>=0.0.9
python-dotenv>=1.2.1

# Data Validation & Settings
pydantic>=2.9.2
pydantic-settings>=2.11.0
email-validator>=2.3.0

# Database
motor>=3.7.1
pymongo>=4.15.3
dnspython>=2.8.0

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt==4.1.3

# Data Science & ML (Base)
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
scikit-learn>=1.5.0
tqdm>=4.66.0
torch>=2.4.0
torchvision>=0.19.0
numba==0.63.1
llvmlite==0.46.0
slicer==0.0.8

# Medical Imaging & Specialized
monai==1.5.1
nibabel>=5.2.0
shap==0.49.1
matplotlib>=3.8.0
icecream==2.1.8
wandb==0.24.0
colorama>=0.4.6
toml>=0.10.0
